RavenDB-18803 - create and modify orchestrator topology endpoints and change cluster observer to have orchestrator nodes handling
/admin/databases/reorder
/admin/databases/promote
PUT admin/orchestrator
DELETE admin/orchestraor
